Move FileNameCaseCollisionsPreventer to utils.py
